Cleanup and LCD working
